+++ src/main/java/com/kaba4cow/futuresscreenerbot/entity/SubscriberSettings.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.entity;
 
-import java.math.BigDecimal;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
@@ -17,15 +15,15 @@ import lombok.Setter;
 public class SubscriberSettings implements SubscriberSettingsProvider {
 
 	@Column(name = "column_pump_threshold")
-	private BigDecimal pumpThreshold;
+	private Double pumpThreshold;
 
 	@Column(name = "column_dump_threshold")
-	private BigDecimal dumpThreshold;
+	private Double dumpThreshold;
 
 	@Column(name = "column_long_liquidation_threshold")
-	private BigDecimal longLiquidationThreshold;
+	private Double longLiquidationThreshold;
 
 	@Column(name = "column_short_liquidation_threshold")
-	private BigDecimal shortLiquidationThreshold;
+	private Double shortLiquidationThreshold;
 
 }
