+++ src/main/java/com/kaba4cow/futuresscreenerbot/properties/screener/PumpScreenerSettingsProperties.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.properties.screener;
 
-import java.math.BigDecimal;
-
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -14,8 +12,8 @@ import lombok.Setter;
 @Component
 public class PumpScreenerSettingsProperties {
 
-	private BigDecimal minPumpThreshold;
+	private Double minPumpThreshold;
 
-	private BigDecimal maxPumpThreshold;
+	private Double maxPumpThreshold;
 
 }
