+++ src/main/java/com/kaba4cow/futuresscreenerbot/properties/screener/ShortLiquidationScreenerSettingsProperties.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.properties.screener;
 
-import java.math.BigDecimal;
-
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -14,8 +12,8 @@ import lombok.Setter;
 @Component
 public class ShortLiquidationScreenerSettingsProperties {
 
-	private BigDecimal minShortLiquidationThreshold;
+	private Double minShortLiquidationThreshold;
 
-	private BigDecimal maxShortLiquidationThreshold;
+	private Double maxShortLiquidationThreshold;
 
 }
