+++ src/main/java/com/kaba4cow/futuresscreenerbot/service/domain/SubscriberFilter.java
@@ -1,7 +1,5 @@
 package com.kaba4cow.futuresscreenerbot.service.domain;
 
-import java.math.BigDecimal;
-
 import org.springframework.stereotype.Component;
 
 import com.kaba4cow.futuresscreenerbot.entity.Event;
@@ -12,8 +10,8 @@ import com.kaba4cow.futuresscreenerbot.service.domain.settingsextractor.Subscrib
 public class SubscriberFilter {
 
 	public boolean filter(Subscriber subscriber, SubscriberSettingsExtractor extractor, Event event) {
-		BigDecimal threshold = extractor.getThreshold(subscriber.getSettings());
-		BigDecimal value = event.getValue();
+		Double threshold = extractor.getThreshold(subscriber.getSettings());
+		Double value = event.getValue();
 		return value.doubleValue() >= threshold.doubleValue();
 	}
